+++ kernel/src/lib.rs
@@ -7,7 +7,7 @@ pub mod uart;
 pub mod common;
 pub mod handler;
 pub mod riscv;
-pub mod memory;
+mod address;
 mod memlayout;
 pub mod vm;
 pub mod timer;
@@ -15,4 +15,4 @@ pub mod syscall;
 pub mod scheduler;
 pub mod capability;
 pub mod object;
-pub mod init_proc;
+pub mod init;
